# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🚗 Auto-École API - Authentification Authentik",
        "description": "API complète pour la gestion de l'authentification avec Authentik comme IAM central. Cette API permet l'inscription, la connexion, la gestion des utilisateurs et des tokens avec un système de contournement du Password Grant Flow.",
        "contact": {
            "name": "Support Auto-École API",
            "url": "https://auto-ecole-api.com",
            "email": "support@auto-ecole-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
...
        },
        {
...
        },
...
    "paths": {
...
        "/api/candidats": {
            "get": {
                "tags": [
                    "👨‍🎓 Candidats"
                ],
                "summary": "📋 Liste de tous les candidats",
                "description": "Récupère la liste paginée de tous les candidats avec leurs informations personnelles",
                "operationId": "getCandidats",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Liste des candidats récupérée"
                    }
                }
            },
            "post": {
                "tags": [
                    "👨‍🎓 Candidats"
                ],
                "summary": "➕ Créer un nouveau candidat",
                "description": "Crée un nouveau candidat avec ses informations personnelles",
                "operationId": "storeCandidat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "personne_id",
                                    "numero_candidat",
                                    "date_naissance",
                                    "lieu_naissance",
                                    "nip",
                                    "type_piece",
                                    "numero_piece",
                                    "nationalite",
                                    "genre"
                                ],
                                "properties": {
                                    "personne_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019a0e34-d153-7330-8cb6-80b14fd8811c"
                                    },
                                    "numero_candidat": {
                                        "type": "string",
                                        "example": "CAN-2025-001"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-15"
                                    },
                                    "lieu_naissance": {
                                        "type": "string",
                                        "example": "Dakar"
                                    },
                                    "nip": {
                                        "type": "string",
                                        "example": "1234567890123"
                                    },
                                    "type_piece": {
                                        "type": "string",
                                        "example": "CNI"
                                    },
                                    "numero_piece": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "nationalite": {
                                        "type": "string",
                                        "example": "Sénégalaise"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "✅ Candidat créé avec succès"
                    },
                    "401": {
                        "description": "❌ Non authentifié"
                    },
                    "422": {
                        "description": "❌ Erreur de validation"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/candidats/complete-profile": {
            "post": {
                "tags": [
                    "👨‍🎓 Candidats"
                ],
                "summary": "✅ Compléter le profil candidat",
                "description": "Permet à un utilisateur connecté de compléter son profil candidat avec ses informations personnelles",
                "operationId": "completeProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date_naissance",
                                    "lieu_naissance",
                                    "nip",
                                    "type_piece",
                                    "numero_piece",
                                    "nationalite",
                                    "genre"
                                ],
                                "properties": {
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-15"
                                    },
                                    "lieu_naissance": {
                                        "type": "string",
                                        "example": "Dakar"
                                    },
                                    "nip": {
                                        "type": "string",
                                        "example": "1234567890123"
                                    },
                                    "type_piece": {
                                        "type": "string",
                                        "example": "CNI"
                                    },
                                    "numero_piece": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "nationalite": {
                                        "type": "string",
                                        "example": "Sénégalaise"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "✅ Profil candidat complété"
                    },
                    "400": {
                        "description": "❌ Profil déjà complété"
                    },
                    "401": {
                        "description": "❌ Non authentifié"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/candidats/mes-dossiers": {
            "get": {
                "tags": [
                    "👨‍🎓 Candidats"
                ],
                "summary": "📁 Mes dossiers",
                "description": "Récupère tous les dossiers du candidat connecté",
                "operationId": "mesDossiers",
                "responses": {
                    "200": {
                        "description": "✅ Liste des dossiers du candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dossiers": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "❌ Profil candidat non complété"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Token d'accès JWT obtenu via l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearer_token": {
                "type": "http",
                "description": "Token d'authentification JWT obtenu via /api/auth/login-direct",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}